Add files via upload
Logic- I have used ORB of OpenCV for feature extraction and the applied clustering logic to classify. 
To run the Application download the docker file Run the following commands in the command file. It Builds and runs docker. and then you can run the Application in it.
To view Code, Please open main.CPP.